+++ Cargo.toml
@@ -6,7 +6,7 @@ members = [
     "shared",
     "testing/conformance",
     "testing/integration",
-    "testing/calibration",
+    "testing/calibration/shared",
     "ipld/*",
     "testing/test_actors",
     "testing/test_actors/actors/*",
@@ -20,10 +20,6 @@ overflow-checks = true
 lto = true
 opt-level = "z"
 #strip = true
-
-# so we don't run out of disk space on Github CI
-[profile.dev.package.fvm_gas_calibration]
-strip = true
 
 [profile.wasm]
 inherits = "release"
