--- testing/calibration/Cargo.toml
@@ -1,45 +1,0 @@
-[package]
-name = "fvm_gas_calibration"
-description = "Filecoin Virtual Machine gas calibration"
-version = "0.1.0"
-edition = "2021"
-license = "MIT OR Apache-2.0"
-authors = ["Protocol Labs", "Filecoin Core Devs"]
-repository = "https://github.com/filecoin-project/ref-fvm"
-publish = false
-
-[dependencies]
-fvm = { version = "3.1.0", path = "../../fvm", default-features = false, features = ["testing", "gas_calibration"] }
-fvm_shared = { version = "3.1.0", path = "../../shared", features = ["testing"] }
-fvm_ipld_car = { version = "0.6.0", path = "../../ipld/car" }
-fvm_ipld_blockstore = { version = "0.1.1", path = "../../ipld/blockstore" }
-fvm_ipld_encoding = { version = "0.3.3", path = "../../ipld/encoding" }
-fvm_integration_tests = { path = "../integration" }
-fvm_gas_calibration_shared = { path = "./shared" }
-actors-v10 = { package = "fil_builtin_actors_bundle", git = "https://github.com/filecoin-project/builtin-actors", branch = "next" }
-
-anyhow = "1.0.47"
-cid = { version = "0.8.5", default-features = false }
-futures = "0.3.19"
-multihash = { version = "0.16.1", default-features = false }
-num-traits = "0.2"
-lazy_static = "1.4.0"
-libsecp256k1 = "0.7.0"
-rand = "0.8.5"
-rand_chacha = "0.3"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-thiserror = "1.0.30"
-bls-signatures = { version = "0.13", default-features = false }
-blake2b_simd = "1.0.0"
-wat = "1.0.52"
-fvm_test_actors = { path = "../test_actors" }
-
-
-[dependencies.wasmtime]
-version = "2.0.2"
-default-features = false
-features = ["cranelift", "parallel-compilation"]
-
-[features]
-default = ["fvm/testing", "fvm_shared/testing"]
